+++ blocksuite/affine/block-root/src/edgeless/gfx-tool/index.ts
@@ -1,6 +1,5 @@
 export { BrushTool } from './brush-tool.js';
 export { ConnectorTool, type ConnectorToolOptions } from './connector-tool.js';
-export { CopilotTool } from './copilot-tool.js';
 export { DefaultTool } from './default-tool.js';
 export { EmptyTool } from './empty-tool.js';
 export { EraserTool } from './eraser-tool.js';
