+++ blocksuite/affine/block-root/src/edgeless/index.ts
@@ -6,4 +6,5 @@ export { EdgelessRootPreviewBlockComponent } from './edgeless-root-preview-block
 export { EdgelessRootService } from './edgeless-root-service.js';
 export * from './gfx-tool';
 export * from './utils/clipboard-utils.js';
+export { sortEdgelessElements } from './utils/clone-utils.js';
 export { isCanvasElement } from './utils/query.js';
