+++ packages/frontend/core/src/blocksuite/presets/ai/widgets/ai-panel/components/state/error.ts
@@ -1,9 +1,6 @@
-import {
-  AIErrorType,
-  type AIItemGroupConfig,
-} from '@blocksuite/affine-components/ai-item';
-import type { EditorHost } from '@blocksuite/block-std';
-import { WithDisposable } from '@blocksuite/global/utils';
+import type { EditorHost } from '@blocksuite/affine/block-std';
+import { AIErrorType, type AIItemGroupConfig } from '@blocksuite/affine/blocks';
+import { WithDisposable } from '@blocksuite/affine/global/utils';
 import { baseTheme } from '@toeverything/theme';
 import { css, html, LitElement, nothing, unsafeCSS } from 'lit';
 import { property } from 'lit/decorators.js';
